+++ app/src/main/kotlin/com/numero/sojodia/ui/timetable/TimeTableRowAdapter.kt
@@ -1,4 +1,4 @@
-package com.numero.sojodia.view.adapter
+package com.numero.sojodia.ui.timetable
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,7 +6,8 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.numero.sojodia.R
 import com.numero.sojodia.model.TimeTableRow
-import com.numero.sojodia.view.TimeTableRowHolder
+import kotlinx.android.extensions.LayoutContainer
+import kotlinx.android.synthetic.main.holder_time_table_row.*
 import java.util.*
 
 class TimeTableRowAdapter : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
@@ -86,4 +87,20 @@ class TimeTableRowAdapter : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
     }
 
     class DescriptionViewHolder(view: View) : RecyclerView.ViewHolder(view)
+
+    class TimeTableRowHolder(override val containerView: View) : RecyclerView.ViewHolder(containerView), LayoutContainer {
+
+        var isCurrentHour: Boolean = false
+            set(value) {
+                field = value
+                rowLayout.isActivated = value
+            }
+
+        fun setTimeTableRow(timeTableRow: TimeTableRow, isNotSchoolTerm: Boolean) {
+            hourTextView.text = "%02d".format(Locale.ENGLISH, timeTableRow.hour)
+            weekdayTextView.text = timeTableRow.createOnWeekdayText(isNotSchoolTerm)
+            saturdayTextView.text = timeTableRow.createOnSaturdayText(isNotSchoolTerm)
+            sundayTextView.text = timeTableRow.createOnSundayText(isNotSchoolTerm)
+        }
+    }
 }
